use_case_2/requirements.txt
openai
ruff
python-dotenv
openpyxl